Redesign implementation
