+++ memmap.md
@@ -1,9 +1,8 @@
 #Memory mapping
-8000 bytes
-0x0000 - 0x1F40
+64000 bytes (64k)
+0x0000 - 0xFFFF
 
-0x1EF0 - 0x1F40 : Screen
-?      - 0x1EEF : Stack
-         0x1C1F (4ppb)
+0xFF5F - 0xFFFF : Screen
+?      - 0xFF5E : Stack
 Program starts at 0x0000
